requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
discord==1.0.1
discord.py==1.3.3
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
multidict==4.7.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
ulid-py==0.0.14
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
